+++ megatron/training/config.py
@@ -114,22 +114,3 @@ class ValidationConfig:
        separate loss for each dataset in the list. This argument requires that no weights are 
        included in the list.
     """
-
-
-@dataclass(kw_only=True)
-class RNGConfig:
-    """Configuration settings for random number generation."""
-
-    seed: int = 1234
-    """Random seed used for python, numpy, pytorch, and cuda."""
-
-    te_rng_tracker: bool = False
-    """Use the Transformer Engine version of the random number generator.
-    Required for CUDA graphs support."""
-
-    inference_rng_tracker: bool = False
-    """Use a random number generator configured for inference."""
-
-    data_parallel_random_init: bool = False
-    """Enable random initialization of params across data parallel ranks"""
-
